+++ .changeset/breezy-spoons-joke.md
@@ -1,0 +1,10 @@
+---
+"guilded.js": minor
+"@guildedjs/rest": minor
+"@guildedjs/webhook-client": minor
+"@guildedjs/ws": minor
+"@guildedjs/gil": minor
+"@guildedjs/guilded-api-typings": minor
+---
+
+Removal of @guildedjs/common
