--- packages/common/README.md
@@ -1,28 +1,0 @@
-# `@guildedjs/common`
-
-[![GitHub](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/guildedjs/guilded.js/blob/main/LICENSE)
-[![npm](https://img.shields.io/npm/v/@guildedjs/common?color=crimson&logo=npm)](https://www.npmjs.com/package/@guildedjs/common)
-[![CI workflows](https://github.com/guildedjs/guilded.js/actions/workflows/ci.yml/badge.svg)](https://github.com/guildedjs/guilded.js/actions/workflows/ci.yml)
-
-Common structures and utilities used in various @guildedjs packages
-
-> ⚠️ If you intend to use `guilded.js` you won't need to install this, as it's already a dependency of that package.
-
-## Installation
-
-You can install this package from [NPM](https://www.npmjs.com/package/@guildedjs/common)
-
--   `npm install @guildedjs/common`
--   `yarn add @guildedjs/common`
-
-## Contributing
-
-Please see the main [README.md](https://github.com/guildedjs/guilded.js) for info on how to contribute to this package or the other `@guildedjs` packages.
-
-## Credits
-
-Concept and structure dervied from https://github.com/discordjs/discord.js/blob/stable/src/structures/MessageEmbed.js#L9 - [Discord.js](https://github.com/discordjs/discord.js) ([License](https://github.com/discordjs/discord.js/blob/stable/LICENSE))
-
-## LICENSE
-
-Licensed under the [MIT License](https://github.com/guildedjs/guilded.js/blob/main/LICENSE)
