--- packages/common/lib/UUID.ts
@@ -1,2 +1,0 @@
-/* istanbul ignore file */
-export { v4 as generateUUID } from "uuid";
