--- packages/common/lib/index.ts
@@ -1,4 +1,0 @@
-/* istanbul ignore file */
-export * from "./CacheCollection";
-export * from "./util";
-export * from "./UUID";
