--- packages/common/package.json
@@ -1,58 +1,0 @@
-{
-    "name": "@guildedjs/common",
-    "version": "2.1.0",
-    "description": "Common structures and utilities used in various @guildedjs packages",
-    "author": "Zaid \"Nico\" <contact@nico.engineer>",
-    "license": "MIT",
-    "types": "types/index.d.ts",
-    "main": "dist/index.js",
-    "typedocMain": "lib/index.ts",
-    "scripts": {
-        "build": "npx tsc",
-        "prepublishOnly": "npx rimraf dist/ && npx rimraf types/ && pnpm run build && npx gen-esm-wrapper . ./dist/index.mjs"
-    },
-    "devDependencies": {
-        "@guildedjs/guilded-api-typings": "workspace:*",
-        "@types/uuid": "8.3.4",
-        "typescript": "4.6.3"
-    },
-    "dependencies": {
-        "@discordjs/collection": "^0.5.0",
-        "uuid": "^8.3.2"
-    },
-    "contributors": [
-        {
-            "name": "Zaid \"Nico\"",
-            "email": "contact@nico.engineer",
-            "url": "https://github.com/zaida04"
-        },
-        {
-            "name": "Skillz4Killz",
-            "email": "guildedjs@drskillz.33mail.com",
-            "url": "https://github.com/Skillz4Killz"
-        },
-        {
-            "name": "Uhuh \"Dylan\"",
-            "email": "dylan@panku.io",
-            "url": "https://github.com/uhuh"
-        }
-    ],
-    "keywords": [
-        "guilded-gg",
-        "guilded-api",
-        "guilded-webhook"
-    ],
-    "files": [
-        "dist",
-        "types"
-    ],
-    "homepage": "https://github.com/guildedjs/guilded.js/tree/main/packages/common#readme",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/guildedjs/guilded.js.git"
-    },
-    "bugs": {
-        "url": "https://github.com/guildedjs/guilded.js/issues"
-    },
-    "gitHead": "eee38a19e0bfa812d7136cc78a6bc99e0b402b0c"
-}
