+++ packages/guilded.js/package.json
@@ -19,7 +19,6 @@
     },
     "dependencies": {
         "@discordjs/collection": "^0.5.0",
-        "@guildedjs/common": "workspace:*",
         "@guildedjs/rest": "workspace:*",
         "@guildedjs/webhook-client": "workspace:*",
         "@guildedjs/ws": "workspace:*"
