+++ packages/rest/package.json
@@ -18,7 +18,6 @@
         "typescript": "4.6.3"
     },
     "dependencies": {
-        "@guildedjs/common": "workspace:*",
         "node-fetch": "^2.6.7",
         "qs": "^6.10.3"
     },
