+++ packages/webhook-client/package.json
@@ -16,7 +16,6 @@
         "typescript": "4.6.3"
     },
     "dependencies": {
-        "@guildedjs/common": "workspace:*",
         "@guildedjs/rest": "workspace:*"
     },
     "contributors": [
