+++ packages/ws/lib/index.ts
@@ -1,4 +1,1 @@
-import WebsocketManager from "./WebSocketManager";
-
 export * from "./WebSocketManager";
-export default WebsocketManager;
