+++ packages/ws/package.json
@@ -15,7 +15,6 @@
         "typescript": "4.6.3"
     },
     "dependencies": {
-        "@guildedjs/common": "workspace:*",
         "@guildedjs/guilded-api-typings": "workspace:*",
         "@types/ws": "8.5.3",
         "typed-emitter": "2.1.0",
